refactor(#367): relocate CrashLog out of wifi_observer/ to diagnostics/
Epic #350 increment. Boot-survival diagnostics were structurally tied to the
WiFi observer, so the roles that need them most -- unattended nRF52 repeaters,
room servers, sensor leaves -- could not use them. The coupling was already
leaking: src/helpers/ui/SSD1306Display.cpp reached into wifi_observer/ from
outside the observer tree.

git mv (history preserved):
  src/helpers/wifi_observer/CrashLog.{h,cpp} -> src/helpers/diagnostics/

Updated all six consumers. Three used a same-directory relative include
("CrashLog.h") which silently broke on the move; those now use the explicit
<helpers/diagnostics/CrashLog.h> form.

BUILD FILTER -- the non-obvious part. build_src_filter enumerates directories
explicitly (+<helpers/*.cpp>, +<helpers/radiolib/*.cpp>, ...), so a file in a
NEW directory is simply never compiled. The first attempt linked with undefined
references to every offband:: symbol because CrashLog.cpp had silently dropped
out of the build. Added +<helpers/diagnostics/*.cpp> at exactly the four sites
that already carry +<helpers/wifi_observer/*.cpp> (heltec_v3, heltec_v4 x2,
xiao_s3_wio), so the SET of envs compiling CrashLog is unchanged -- no env
starts or stops building it.

Scope guard: NO behavior change. CrashLog still depends on WifiObserverConfig.h
(now included by full path, with a note); breaking that dependency and
abstracting the storage backend for nRF52 are the next increments.

Builds: heltec_v4_companion_observer_wifi (heaviest consumer) + RAK_3401_repeater
-- both SUCCESS.
