+++ variants/heltec_v3/platformio.ini
@@ -439,3 +439,4 @@ build_src_filter =
   ${env:Heltec_v3_companion_radio_ble.build_src_filter}
   +<helpers/wifi_observer/*.cpp>
   +<helpers/config/*.cpp>            ; #364: shared role-agnostic config dispatch
+  +<helpers/diagnostics/*.cpp>   ; #350: CrashLog moved out of wifi_observer/
