+++ variants/heltec_v4/platformio.ini
@@ -585,6 +585,7 @@ build_src_filter =
   ${env:heltec_v4_companion_radio_ble.build_src_filter}
   +<helpers/wifi_observer/*.cpp>
   +<helpers/config/*.cpp>            ; #364: shared role-agnostic config dispatch
+  +<helpers/diagnostics/*.cpp>   ; #350: CrashLog moved out of wifi_observer/
 [env:heltec_v4_kiss_modem]
 extends = Heltec_lora32_v4
 build_src_filter = ${Heltec_lora32_v4.build_src_filter}
@@ -615,6 +616,7 @@ build_src_filter =
   ${env:heltec_v4_tft_companion_radio_ble.build_src_filter}
   +<helpers/wifi_observer/*.cpp>
   +<helpers/config/*.cpp>            ; #364: shared role-agnostic config dispatch
+  +<helpers/diagnostics/*.cpp>   ; #350: CrashLog moved out of wifi_observer/
 lib_deps =
   ${env:heltec_v4_tft_companion_radio_ble.lib_deps}
   ${esp32_ble.lib_deps}
