+++ variants/xiao_s3_wio/platformio.ini
@@ -273,6 +273,7 @@ build_src_filter =
   ${env:Xiao_S3_WIO_companion_radio_ble.build_src_filter}
   +<helpers/wifi_observer/*.cpp>
   +<helpers/config/*.cpp>            ; #364: shared role-agnostic config dispatch
+  +<helpers/diagnostics/*.cpp>   ; #350: CrashLog moved out of wifi_observer/
 [env:Xiao_S3_WIO_kiss_modem]
 extends = Xiao_S3_WIO
 build_src_filter = ${Xiao_S3_WIO.build_src_filter}
